Pushing the "Develop" files for the Module 1 Challenge: Code Refactor
I have already done the bulk of refactoring with this push. I was
having some issues with Git not recognizing my key even though I had
made a demo practice in prep for making a new repo for this project.
Now that I spent a good chunk of time figuring it out, I am pushing the
work I have done so far.

This push includes technical improvements to the code: I fixed
the broken link, consolidated and organized the CSS, and added various
CSS comments including a table of contents and sections to improve the
readability of the stylesheet.